+++ ICD.Connect.Protocol.Network/Ports/Web/HttpPort.SimplSharp.cs
@@ -235,6 +235,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 
 			Accept = SOAP_ACCEPT;
 			m_HttpsClient.IncludeHeaders = false;
+			response = null;
 
 			m_ClientBusySection.Enter();
 
@@ -270,10 +271,17 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 					return Dispatch(request, out response);
 				}
 			}
+			catch (Exception e)
+			{
+				Log(eSeverity.Error, "Failed to dispatch SOAP - {0}", e.Message);
+			}
 			finally
 			{
 				m_ClientBusySection.Leave();
 			}
+
+			SetLastRequestSucceeded(false);
+			return false;
 		}
 
 		#endregion