+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Fixed case where HTTP/S port would not go offline if a SOAP request fails to dispatch
+
 ## [8.2.3] - 2019-06-04
 ### Changed
  - Doubly-escaping HTTP/S paths for Crestron due to path encodings being stripped